requirements.txt
appdirs==1.4.4
argon2==0.1.10
argon2-cffi==20.1.0
asgiref==3.3.4
astroid==2.6.2
attrs==21.2.0
black==21.6b0
certifi==2021.5.30
cffi==1.14.6
click==8.0.1
Django==3.2.4
djangorestframework==3.12.4
iniconfig==1.1.1
isort==5.9.2
lazy-object-proxy==1.6.0
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.0
pathspec==0.8.1
pluggy==0.13.1
psycopg2==2.9.1
py==1.10.0
pycparser==2.20
pylint==2.9.3
pyparsing==2.4.7
pytest==6.2.4
python-dotenv==0.18.0
pytz==2021.1
regex==2021.7.6
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1
